Work in progress Food logging

+++ FitnessDAL/DTO/MealLogDTO.cs
@@ -9,9 +9,9 @@ namespace FitnessDAL.DTO
         public int UserID { get; set; }
         public DateOnly Date { get; set; }
         public int TotalCalories { get; set; }
-        public int TotalProtein { get; set; }
-        public int TotalCarbohydrates { get; set; }
-        public int TotalFat { get; set; }
+        public decimal TotalProtein { get; set; }
+        public decimal TotalCarbohydrates { get; set; }
+        public decimal TotalFat { get; set; }
         public List<MealDTO> Meals { get; set; } = new();
     }
 }
